+++ src/Http/Middleware/ChronicleUiEnabled.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Chronicle\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+class ChronicleUiEnabled
+{
+    public function handle(Request $request, Closure $next): Response
+    {
+        if (! config('chronicle.ui.enabled', false)) {
+            abort(404);
+        }
+
+        return $next($request);
+    }
+}